icmp extraction "fixed"
I'm not sure if anyone has done any real testing with hping's icmp extraction capabilities, but the changes I've made now make it more functional than the previous implementation.  It still seems to produce files larger than the originals (hence why I will not be implementing https://github.com/GinjaChris/pentmenu/issues/35) however the fies are usable.  As an exampe, a transferred PDF file was twice the original size, and contained the original data *repeated twice*, but opened correctly as a PDF file.  
An extracted text file would not open with a text editor, but did open correctly in LibreOffice Writer.

I believe these issues are due to the way that hping sends the files over ICMP; at the very least it adds extra 'nonsense' to the end of the files since it continues to send the ICMP packets even after the file has been transferred.

On the other hand I could just be doing something incredibly silly.

Anyway, the ICMP extraction module DOES work, it's jst a little unorthodix in the results.